pyproject.toml
[tool.poetry]
name = "drf-course-work"
version = "0.1.0"
description = ""
authors = ["alexander anciferov <sania.antsiferoff@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.7"
psycopg2-binary = "^2.9.10"
pillow = "^11.1.0"
djangorestframework = "^3.15.2"
python-dotenv = "^1.0.1"
djangorestframework-simplejwt = "^5.5.0"
django-phonenumber-field = {extras = ["phonenumbers"], version = "^8.0.0"}
black = "^25.1.0"
requests = "^2.32.3"
pytz = "^2025.1"
celery = "^5.4.0"
django-celery-beat = "^2.7.0"
redis = "^5.2.1"
drf-yasg = "^1.21.10"
django-cors-headers = "^4.7.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
